+++ database/seeders/DatabaseSeeder.php
@@ -12,6 +12,8 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         $this->call([
+            RoleSeeder::class,
+            PengaturanSeeder::class,
             UserSeeder::class,
             FaqSeeder::class,
         ]);